backend/requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
psycopg2-binary
pydantic
